Multiple fixes. See below.
* Moved `registerColored` to the proxy classes.
* Added mcjty's compatibility layer to Gradle
* Removed unnecessary/broken imports
